Revert "Introduce Dlang codegen."
This reverts commit 8df0ed673737a32a227533743b0cf4b0d4c7c528.

Reason for revert: D backend is not ready yet: there are no examples
and tests, the work on it has stopped. In the future support for D and
other languages should be added in a generic way via language syntax
files, rather than modifying re2c codegen to support each new language.