Wrap lexer body in braces only if there're local variable declarations.
This is a revert of commit 77714fce9463abfe4420af93e27bb8ac4c24719c.

While it is easier to always generate braces, in some languages (Go) it
is a compilation error to jump into the wrapped block (a scenario that
happens if there is an outer state switch generated with `getstate:re2c`
directive.

This regression was found because one of the examples for Go didn't
compile (examples/state/push.go). Now it does.