In rec/func mode, don't elide transitions in "if/else-if/else" branches.
This is because in functional languages like OCaml "if/else-if/else" is
an expression where all branches must have the same type (except perhaps
for a single "if" branch that may have void type.

One exception is the case of a single branch that goes to the next state
(in that case "if/else-if/else" construct is not needed).