Change YYFN format to allow type modifiers.
In some languages type modifiers go on the left of function parameters,
while the type itself is on the right (like `mut n: usize` in Rust). For
such cases, we need to allow optional third component of an argument
descriptor in YYFN. So, change its format from a list of 2*N strings to
a list of N strings with semicolon-separated parts (name, type, optional
modifiers).

Add `fn:sep` configuration to allow the users configure the separator.

Precompute function parameters and arguments, as they are common for all
function definitions / calls.