Add configurations `monadic` and `variable:yyfill`.
The new configurations are needed to support monadic lexers in the
upcoming Haskell backend.

Configuration `monadic` affects control flow constructs: when it is on,
the "do" keyword is generated before nested blocks.

Configuration `variable:yyfill" allows one to decouple YYFILL invocation
from the check whether YYFILL has failed. This is needed becausse YYFILL
performs I/O (like reading from file), so it should return the updated
state of the world together with the status whether it has failed or not.