Add syntax configurations for tag operations.
The following configurations have been added:
    - code:shiftstag
    - code:shiftmtag
    - code:stagp
    - code:stagn
    - code:mtagp
    - code:mtagn
    - code:copystag
    - code:copymtag

Mark `code:shiftstag` and `code:shiftmtag` as oneline configurations
(in fact they should be named one-statement, not oneline).

Note that the user can define YYSHIFTSTAG/YYSHIFTMTAG as multi-statement
without the curly braces, and thus break the assumption re2c makes about
the generated code. But it is also the case with YYFILL (and has always
been), so we don't make the situation much worse. This requirement
should be documented, though.