Add API primitives SETACCEPT/GETACCEPT and configurations `code:yysetaccept`/`code:yygetaccept`.
The new primitives and configurations allow one to override `yyaccept`
syntax for a single source file or for a given language backend.

For existing languages the new primitives have default definitions:
nothing changes in the code generation, and no new definitions are
required from the user. The change is fully backwards compatible.