Always apply default syntax file, but allow overriding it.
This way the user may provide a syntax file with only a few overrides
instead of copy-pasting default definitions for all configurations.

As an example, add a test for the Haskell backend that pretends to be a
test for the Python backend, but overrides default Python config with
Haskell config. We don't use default C config in this case as it
defines a few optional configurations for oneline switch cases/ifs that
are not overridden by the Haskell config. The generated code is
identical to examples/haskell/01_basic.re.

Also, remove a stale TODO comment.