Allow setting immutable configurations in syntax files.
This fixes a problem with libre2c introduced by the previous commit
8e9bfda737ffd09590e23474432341f2d2a53ca3 which required that all code options
are defined (non-null). The problem was caused by testing if `line_dirs` option
is defined: the test happened in the code that is shared with libre2c, while
normally such tests happen in the code that is not shared. This commit moves
`line_dirs` initialization to config files, so the faulty test is removed
entirely.

It also changes the way that implied imuutable options are set: former
`set_default_*` functions that soft-set an option (mark it as not user-defined)
are replaced with `set_*` functions that hard-set the option (mark it as
user-defined). It shouldn't matter for immutable options, as they cannot be
redefined by the user.

an option  only set an option specified in a syntax
file if it's not explicitly defined by the user. No tests have changed.